arch-power: Fix remote GDB semantics
This fixes the endianness conversions applied to the integer
registers when reading or setting them. This also changes the
registers widths to 64 bits wherever required for compliance
with the 64-bit Power architecture.

The format of the 'g' packet that GDB expects for targets of
type "powerpc:common64" is also fixed by maintaining correct
offsets for the registers and introducing additional padding
to keep the correct packet size.

Change-Id: I7c64c0a2e355a2e5fbab097e9d24c773c475db85
Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>