arch-power: Fix special-purpose register definitions
This converts the definitions of all the currently defined
Special Purpose Registers (SPRs) to miscellaneous registers
which allows us to print the corresponding SPR name in the
debug logs rather than an ambiguous register number making
things much easier to correlate.

This also fixes disassembly generation for the move to and
from instructions corresponding to some of the SPRs.

Change-Id: I4c477e886adac5761cd9e9228bcdebd958ee22eb
Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>