arch-power: sim-se: Fix symbol resolution for big endian mode
The ELF ABI used for big-endian binaries specifies that the
symbol corresponding to a function will actually point to a
function descriptor (in the ".opd" section) rather than the
actual entry point of the function (in the ".text" section).
A descriptor comprises three doublewords of which the first
doubleword contains the actual entry point address.

To make symbol resolution seamless, this rebuilds the symbol
table in a way such that the symbol names are mapped to the
function's entry point address rather than the descriptor's
address.

Change-Id: I392240069b72f52487bab641f7a81682f34cecd4
Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>