configs: Fix waiting on remote debugger
Commit 2c75e58cac ("sim,cpu: Move the remote GDB stub into
the workload.") moved "wait_for_gdb" to the Workload class
which breaks se.py as it continues to rely on that property
being present in BaseCPU. This makes sure that an instance
of Workload is used instead.

Also, owing to its boolean nature, the parameter should
ideally not expect any additional values. Hence, it is
associated with the "store_true" action.

Change-Id: I4a00b29d283df36ebf833c9125651cd6deb52a4f
Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>